Move UncaughtExceptionGuard into workflow-testing
