--- internal-testing-utils/api/internal-testing-utils.api
@@ -1,10 +1,0 @@
-public final class com/squareup/workflow1/internal/util/UncaughtExceptionGuard {
-	public fun <init> ()V
-	public final fun reportUncaught (Ljava/lang/Throwable;)V
-	public final fun runRethrowingUncaught (Lkotlin/jvm/functions/Function0;)Ljava/lang/Object;
-}
-
-public final class com/squareup/workflow1/internal/util/UncaughtExceptionGuardKt {
-	public static final fun rethrowingUncaughtExceptions (Lkotlin/jvm/functions/Function0;)V
-}
-
