--- internal-testing-utils/dependencies/runtimeClasspath.txt
@@ -1,5 +1,0 @@
-org.jetbrains.kotlin:kotlin-bom:2.0.21
-org.jetbrains.kotlin:kotlin-stdlib-jdk7:2.0.21
-org.jetbrains.kotlin:kotlin-stdlib-jdk8:2.0.21
-org.jetbrains.kotlin:kotlin-stdlib:2.0.21
-org.jetbrains:annotations:13.0
