--- internal-testing-utils/gradle.properties
@@ -1,3 +1,0 @@
-POM_ARTIFACT_ID=workflow-internal-testing-utils
-POM_NAME=Workflow internal testing utilities
-POM_PACKAGING=jar
