+++ workflow-testing/api/workflow-testing.api
@@ -102,6 +102,16 @@ public final class com/squareup/workflow1/testing/RenderTesterWorkersKt {
 	public static final fun keyDescription (Ljava/lang/String;)Ljava/lang/String;
 }
 
+public final class com/squareup/workflow1/testing/UncaughtExceptionGuard {
+	public fun <init> ()V
+	public final fun reportUncaught (Ljava/lang/Throwable;)V
+	public final fun runRethrowingUncaught (Lkotlin/jvm/functions/Function0;)Ljava/lang/Object;
+}
+
+public final class com/squareup/workflow1/testing/UncaughtExceptionGuardKt {
+	public static final fun rethrowingUncaughtExceptions (Lkotlin/jvm/functions/Function0;)V
+}
+
 public final class com/squareup/workflow1/testing/WorkerSink : com/squareup/workflow1/Worker {
 	public fun <init> (Ljava/lang/String;Lkotlin/reflect/KClass;)V
 	public fun doesSameWorkAs (Lcom/squareup/workflow1/Worker;)Z
