+++ workflow-testing/src/main/java/com/squareup/workflow1/testing/WorkflowTestRuntime.kt
@@ -11,7 +11,6 @@ import com.squareup.workflow1.Workflow
 import com.squareup.workflow1.WorkflowInterceptor
 import com.squareup.workflow1.WorkflowInterceptor.WorkflowSession
 import com.squareup.workflow1.config.JvmTestRuntimeConfigTools
-import com.squareup.workflow1.internal.util.UncaughtExceptionGuard
 import com.squareup.workflow1.renderWorkflowIn
 import com.squareup.workflow1.testing.WorkflowTestParams.StartMode.StartFresh
 import com.squareup.workflow1.testing.WorkflowTestParams.StartMode.StartFromCompleteSnapshot
